Update to v2024.11.2
A security patch has just been released, update to this version.
Add some other modifications like removing stripe, braintree and paypal keys.

+++ scripts/apply_patches.sh
@@ -57,7 +57,9 @@ replace_embedded_svg_icon \
 
 echo "Remove non-free bitwarden_license/ code"
 rm -rf ./bitwarden_license/
-rm -rf ./apps/web/src/app/tools/access-intelligence/
+if [ -d "./apps/web/src/app/tools/access-intelligence/" ]; then
+    rm -rf ./apps/web/src/app/tools/access-intelligence/
+fi
 
 echo "Using patch: ${PATCH_NAME}"
 git apply "../patches/${PATCH_NAME}" --reject
